Precompute alignment for injections and time-frequency.
The computation of Nperseg, Npersplice, nsegs, nsplices etc
have been moved to the argument parsing. This allows us to
immediately tell if a run is going to fail due to misalignment
either between h1 and h1_gw in LISA.makedata(), or in the last
chunk in tser2fser(), saving precious user time.

This hopefully solves codeberg issue #7.