Disallow AET external data in LDC format.
Problem: BLIP uses the Adams & Cornish convention for defining AET,
as can be seen in _process_mldc_data().

This is not the convention in DDPC's LISA Rosetta Stone which
I use. So it's not useful for me to give AET data to BLIP; only
confusion can arise (and has arisen).